srcs/backend/history/requirements.txt
django
django-rest-framework
psycopg2-binary
django-watchman
django-cors-headers
django-prometheus
